Run the embedder in-process instead of spawning a child process
   Previously the root `formal-web` binary spawned `formal-web-embedder` as a
   separate process, forwarding CLI arguments via command-line serialization.
   This commit eliminates that indirection by making the embedder crate a
   library (`embedder`) that the root binary depends on directly.

   Changes:
   - Created `embedder/src/lib.rs` with the runtime logic
     (`run_app_with_options`, `run_default`, `run_webdriver`, `run_cdp`) and
     re-exports from the `event_loop` module.
   - Added `[lib]` section to `embedder/Cargo.toml` so the crate produces both
     a library and a standalone binary.
   - Thinned `embedder/src/main.rs` to a CLI parser that delegates to the library.
   - Removed `run_embedder_process` and all process-spawning helpers from
     `src/main.rs` (~100 lines removed). Now calls `embedder::run_default()`,
     `embedder::run_webdriver()`, etc. directly.
   - Added `embedder` dependency to root `Cargo.toml`.
   - Updated `AGENTS.md` to reflect the new architecture.
   - Added process architecture section to readme
