+++ src/main.rs
@@ -1,7 +1,7 @@
 use clap::{Parser, Subcommand};
 use log::error;
 use std::ffi::OsString;
-use std::process::{Command as ProcessCommand, ExitCode, Stdio};
+use std::process::ExitCode;
 use verification::run_validation_from_iter;
 
 #[derive(Parser, Debug)]
@@ -51,105 +51,6 @@ fn delegated_tla_validate_command() -> Option<ExitCode> {
     })
 }
 
-fn run_embedder_process(embedder_args: Vec<OsString>) -> Result<(), String> {
-    // The embedder is a workspace member, so cargo build --release already
-    // produces target/{profile}/formal-web-embedder alongside the root
-    // binary.  Find it and run it directly rather than spawning a nested
-    // cargo invocation (which would deadlock on the workspace target dir).
-    let current_exe = std::env::current_exe()
-        .map_err(|error| format!("failed to resolve current executable: {error}"))?;
-    let embedder_path = current_exe
-        .parent()
-        .ok_or_else(|| String::from("failed to resolve executable directory"))?
-        .join("formal-web-embedder");
-
-    if !embedder_path.is_file() {
-        return Err(format!(
-            "embedder binary not found at {}; run `cargo build --release -p embedder --bin formal-web-embedder` first",
-            embedder_path.display()
-        ));
-    }
-
-    let status = ProcessCommand::new(&embedder_path)
-        .args(&embedder_args)
-        .stdin(Stdio::inherit())
-        .stdout(Stdio::inherit())
-        .stderr(Stdio::inherit())
-        .status()
-        .map_err(|error| {
-            format!(
-                "failed to execute embedder at {}: {error}",
-                embedder_path.display()
-            )
-        })?;
-
-    if status.success() {
-        Ok(())
-    } else {
-        Err(format!(
-            "embedder process exited with status {}",
-            status
-                .code()
-                .map(|code| code.to_string())
-                .unwrap_or_else(|| String::from("unknown"))
-        ))
-    }
-}
-
-fn run_embedder_default(verify: bool, headless: bool) -> Result<(), String> {
-    let mut args = Vec::<OsString>::new();
-    if verify {
-        args.push(OsString::from("--verify"));
-    }
-    if headless {
-        args.push(OsString::from("--headless"));
-    }
-    run_embedder_process(args)
-}
-
-fn run_embedder_webdriver(verify: bool, args: automation::WebDriverArgs) -> Result<(), String> {
-    let mut embedder_args = Vec::<OsString>::new();
-    if verify {
-        embedder_args.push(OsString::from("--verify"));
-    }
-    embedder_args.push(OsString::from("webdriver"));
-    embedder_args.push(OsString::from("--port"));
-    embedder_args.push(OsString::from(args.port.to_string()));
-    if let Some(cdp_port) = args.cdp_port {
-        embedder_args.push(OsString::from("--cdp-port"));
-        embedder_args.push(OsString::from(cdp_port.to_string()));
-    }
-    if args.headless {
-        embedder_args.push(OsString::from("--headless"));
-    }
-    if let Some(startup_url) = args.startup_url {
-        embedder_args.push(OsString::from("--startup-url"));
-        embedder_args.push(OsString::from(startup_url));
-    }
-    if args.exit_on_session_delete {
-        embedder_args.push(OsString::from("--exit-on-session-delete"));
-    }
-    run_embedder_process(embedder_args)
-}
-
-fn run_embedder_cdp(verify: bool, args: automation::CdpArgs) -> Result<(), String> {
-    let mut embedder_args = Vec::<OsString>::new();
-    if verify {
-        embedder_args.push(OsString::from("--verify"));
-    }
-    embedder_args.push(OsString::from("cdp"));
-    embedder_args.push(OsString::from("--port"));
-    embedder_args.push(OsString::from(args.port.to_string()));
-    if args.headless {
-        embedder_args.push(OsString::from("--headless"));
-    }
-    if let Some(startup_url) = args.startup_url {
-        embedder_args.push(OsString::from("--startup-url"));
-        embedder_args.push(OsString::from(startup_url));
-    }
-    run_embedder_process(embedder_args)
-}
-
 fn main() -> ExitCode {
     env_logger::init();
     if let Some(exit_code) = delegated_tla_validate_command() {
@@ -173,9 +74,11 @@ fn main() -> ExitCode {
     }
 
     let result = match command {
-        None => run_embedder_default(cli.verify, cli.headless),
-        Some(CommandKind::WebDriver(args)) => run_embedder_webdriver(cli.verify, args),
-        Some(CommandKind::Cdp(args)) => run_embedder_cdp(cli.verify, args),
+        None => embedder::run_default(cli.verify, cli.headless),
+        Some(CommandKind::WebDriver(args)) => {
+            embedder::run_webdriver(args, cli.verify, cli.headless)
+        }
+        Some(CommandKind::Cdp(args)) => embedder::run_cdp(args, cli.verify, cli.headless),
         Some(CommandKind::Wpt { .. }) => Ok(()),
     };
